Bugfixing.
When trying to create a target angle via scaling by percentage this will create uneven results when the rectangle to be scales is much longer than it is wide. I added some basic trig math to try and determine how much each axis should be scaled first outward then inward. The result is a much more uniform angle around the chip. There is still an error with the logic as it seems like the tangent math node expects radians but when I convert the angle to radians the results are really bad. So I just eyeballed it since this is just minor detailing. I will revisit at some point in the future since my OCD demands that it be perfect.